yant/org-set-preceding-task: Trigger NEXT, but not SCHEDULED
It is not always convenient to trigger scheduled today for a chain of
tasks - it is not a given that I really want to continue the
subsequent task immediately after finishing previous one in the chain.
NEXT is good enough to show up with lower priority in the third layer
of agenda instead of showing up immediately.